chore: more file refactors
